Match-id-63975ad75e6352f636b25975b61ac7625b3269a0

+++ ait_backend/llm/dump/atb_probe.cpp
@@ -358,7 +358,8 @@ namespace Mki {
                 return sign * 0.0f; // Zero
             } else {
                 // Unnormalized number
-                float result = sign * std::ldexp(static_cast<float>(mantissa), 1 - exponentBias - mantissaBits);
+                float result = sign * std::ldexp(static_cast<float>(mantissa),
+                1 - static_cast<int>(exponentBias) - static_cast<int>(mantissaBits));
                 return result;
             }
         }
